backend/board_app/api/permissions.py
from rest_framework.permissions import BasePermission


class IsMemberOrOwner(BasePermission):
    def has_permission(self, request, view):
        return request.user and request.user.is_authenticated

    def has_object_permission(self, request, view, obj):
        user = request.user
        return user == obj.owner or obj.members.filter(id=user.id).exists()


class IsOwner(BasePermission):
    def has_object_permission(self, request, view, obj):
        return request.user == obj.owner
